requirements.txt
Django>=6.0a1,<7.0
django-environ>=0.12
psycopg[binary]>=3.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
django-summernote>=0.8.20
boto3>=1.34
